Delete Inventory & selectMenu func move to Main & inventory func move to VendingMachine & Edited Singleton
